stack/init/requirements.txt
psycopg2-binary
pandas
sqlalchemy
pymongo
dotenv